refactor: move code to controlflowbased objective function
